backend/requirements.txt
Django==3.2.3
django-filter==23.2
djangorestframework==3.12.4
flake8==6.0.0
django-colorfield
psycopg2-binary==2.9.3 
djoser==2.1.0
sorl-thumbnail==12.9.0
Pillow==9.0.0
# drf-extra-fields==3.7.0